+++ CHANGES.rst
@@ -132,6 +132,8 @@ Changelog
 - Enable 'Allow other value' on selection field using select2 widget [quang]
 - Optionally don't display results on initial search form load. New setting [quang]
 - Display DB name in IDE title bar
+- Update email validation to be RFC5322 compliant [instification]
+
 
 2.0b4 (2016-05-27)
 ------------------